Move gas model to a separate category and integrate into docs
