+++ docs/source/stonesoup.models.measurement.rst
@@ -24,3 +24,9 @@ Categorical
 .. automodule:: stonesoup.models.measurement.categorical
     :show-inheritance:
     :inherited-members:
+
+Gas
+---
+.. automodule:: stonesoup.models.measurement.gas
+    :show-inheritance:
+    :inherited-members:
